+++ mininterface/__init__.py
@@ -1,4 +1,3 @@
-
 from ._lib.run import run
 from ._mininterface import Mininterface
 from .exceptions import Cancelled